refs #1640 fixed additional size_t issues in develop

+++ include/qore/intern/qore_qf_private.h
@@ -307,7 +307,7 @@ struct qore_qf_private {
       if (rc > 0)
 	 do_write_event_unlocked(rc, rc, len);
       else if (xsink && rc < 0)
-         xsink->raiseErrnoException("FILE-WRITE-ERROR", errno, "failed writing " QLLD " byte%s to File", len, len == 1 ? "" : "s");
+         xsink->raiseErrnoException("FILE-WRITE-ERROR", errno, "failed writing " QLLD " byte%s to File", (int64)len, len == 1 ? "" : "s");
 
       return rc;
    }
@@ -395,7 +395,7 @@ struct qore_qf_private {
 	    if (rc >= 0)
 	       break;
             if (errno != EINTR) {
-               xsink->raiseErrnoException("FILE-READ-ERROR", errno, "error reading file after " QLLD " bytes read in ReadOnlyFile::%s()", br, mname);
+               xsink->raiseErrnoException("FILE-READ-ERROR", errno, "error reading file after " QLLD " bytes read in ReadOnlyFile::%s()", (int64)br, mname);
                break;
             }
 	 }
